reorganize detectors
