+++ src/steelhead_gazebo/steelhead_gazebo/bounding_box_image_saver.py
