+++ src/steelhead_gazebo/steelhead_gazebo/clean_data.py
