+++ src/steelhead_object_recognition/CMakeLists.txt
@@ -1,87 +1,30 @@
 cmake_minimum_required(VERSION 3.5)
 project(steelhead_object_recognition)
 
-# Default to C99 for C and C++14 for C++
-if(NOT CMAKE_C_STANDARD)
-  set(CMAKE_C_STANDARD 99)
-endif()
+# Default to C++14
 if(NOT CMAKE_CXX_STANDARD)
   set(CMAKE_CXX_STANDARD 14)
 endif()
 
-# Compiler options
-if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
-  add_compile_options(-Wall -Wextra -Wpedantic)
-endif()
-
 # Find dependencies
 find_package(ament_cmake REQUIRED)
-find_package(rclcpp REQUIRED)
-find_package(rclcpp_components REQUIRED)
-find_package(std_msgs REQUIRED)
-find_package(steelhead_interfaces REQUIRED)
-find_package(cv_bridge REQUIRED)
-find_package(sensor_msgs REQUIRED)
-find_package(image_transport REQUIRED)
-find_package(Boost REQUIRED COMPONENTS filesystem)
+find_package(ament_cmake_python REQUIRED)
+find_package(rclpy REQUIRED)
 
-# Include directories
-include_directories(
-  include
-  ${OpenCV_INCLUDE_DIRS}
-  ${Boost_INCLUDE_DIRS}
-)
+# Install Python module package
+ament_python_install_package(${PROJECT_NAME})
 
-# Create a shared library
-add_library(steelhead_object_recognition SHARED
-  src/object_recognizer.cpp
+install(PROGRAMS
+  steelhead_object_recognition/yolo_detector.py
+  DESTINATION lib/${PROJECT_NAME}
 )
-
-# Link libraries
-target_link_libraries(steelhead_object_recognition
-  ${OpenCV_LIBRARIES}
-  ${Boost_LIBRARIES}
-)
-
-# Register component nodes
-rclcpp_components_register_nodes(steelhead_object_recognition 
-  "steelhead_object_recognition::ObjectRecognizer"
-)
-
-# Define target dependencies explicitly
-ament_target_dependencies(steelhead_object_recognition
-  rclcpp
-  rclcpp_components
-  std_msgs
-  steelhead_interfaces
-  cv_bridge
-  sensor_msgs
-  image_transport
-  OpenCV
-  Boost
-)
-
-# Install the library and other directories
-install(TARGETS
-  steelhead_object_recognition
-  ARCHIVE DESTINATION lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
-)
+ament_export_dependencies(ament_cmake_python)
 
 install(DIRECTORY
   config
   launch
   DESTINATION share/${PROJECT_NAME}/
 )
-
-# Testing
-if(BUILD_TESTING)
-  find_package(launch_testing_ament_cmake REQUIRED)
-  add_launch_test(
-    test/test_object_recognition_launch.py
-  )
-endif()
 
 # Package configuration
 ament_package()
