--- src/steelhead_object_recognition/config/tiny_yolov3.yaml
@@ -1,97 +1,0 @@
-#There is an issue with ComposableNode where it doesn't understand the typical nested YAML structure
-#https://github.com/ros2/rclcpp/issues/715
-
-weights_filename: "tiny.weights"
-cfg_filename: "yolov3-tiny.cfg"
-weights_url: "https://pjreddie.com/media/files/tiny.weights"
-cfg_url: "https://raw.githubusercontent.com/pjreddie/darknet/master/cfg/yolov3-tiny.cfg"
-backend: 3 # cv::dnn::DNN_BACKEND_OPENCV
-target: 0 # cv::dnn::DNN_TARGET_CPU
-conf_threshold: 0.2
-nms_threshold: 0.2
-scale: 0.00392
-inp_width: 416
-inp_height: 416
-swap_rb: true
-mean: 0.5
-classes:
-  - "person"
-  - "bicycle"
-  - "car"
-  - "motorbike"
-  - "aeroplane"
-  - "bus"
-  - "train"
-  - "truck"
-  - "boat"
-  - "traffic light"
-  - "fire hydrant"
-  - "stop sign"
-  - "parking meter"
-  - "bench"
-  - "bird"
-  - "cat"
-  - "dog"
-  - "horse"
-  - "sheep"
-  - "cow"
-  - "elephant"
-  - "bear"
-  - "zebra"
-  - "giraffe"
-  - "backpack"
-  - "umbrella"
-  - "handbag"
-  - "tie"
-  - "suitcase"
-  - "frisbee"
-  - "skis"
-  - "snowboard"
-  - "sports ball"
-  - "kite"
-  - "baseball bat"
-  - "baseball glove"
-  - "skateboard"
-  - "surfboard"
-  - "tennis racket"
-  - "bottle"
-  - "wine glass"
-  - "cup"
-  - "fork"
-  - "knife"
-  - "spoon"
-  - "bowl"
-  - "banana"
-  - "apple"
-  - "sandwich"
-  - "orange"
-  - "broccoli"
-  - "carrot"
-  - "hot dog"
-  - "pizza"
-  - "donut"
-  - "cake"
-  - "chair"
-  - "sofa"
-  - "pottedplant"
-  - "bed"
-  - "diningtable"
-  - "toilet"
-  - "tvmonitor"
-  - "laptop"
-  - "mouse"
-  - "remote"
-  - "keyboard"
-  - "cell phone"
-  - "microwave"
-  - "oven"
-  - "toaster"
-  - "sink"
-  - "refrigerator"
-  - "book"
-  - "clock"
-  - "vase"
-  - "scissors"
-  - "teddy bear"
-  - "hair drier"
-  - "toothbrush"
