--- src/steelhead_object_recognition/config/tiny_yolov4_pipeline.yaml
@@ -1,96 +1,0 @@
-object_recognizer:
-  ros__parameters:
-    weights_filename: "yolov4-tiny.weights"
-    cfg_filename: "yolov4-tiny.cfg"
-    weights_url: "https://github.com/AlexeyAB/darknet/releases/download/darknet_yolo_v4_pre/yolov4-tiny.weights"
-    cfg_url: "https://raw.githubusercontent.com/AlexeyAB/darknet/master/cfg/yolov4-tiny.cfg"
-    backend: 3 # cv::dnn::DNN_BACKEND_OPENCV
-    target: 0 # cv::dnn::DNN_TARGET_CPU
-    conf_threshold: 0.2
-    nms_threshold: 0.2
-    scale: 0.00392
-    inp_width: 416
-    inp_height: 416
-    swap_rb: true
-    mean: 0.5
-    classes:
-      - "person"
-      - "bicycle"
-      - "car"
-      - "motorbike"
-      - "aeroplane"
-      - "bus"
-      - "train"
-      - "truck"
-      - "boat"
-      - "traffic light"
-      - "fire hydrant"
-      - "stop sign"
-      - "parking meter"
-      - "bench"
-      - "bird"
-      - "cat"
-      - "dog"
-      - "horse"
-      - "sheep"
-      - "cow"
-      - "elephant"
-      - "bear"
-      - "zebra"
-      - "giraffe"
-      - "backpack"
-      - "umbrella"
-      - "handbag"
-      - "tie"
-      - "suitcase"
-      - "frisbee"
-      - "skis"
-      - "snowboard"
-      - "sports ball"
-      - "kite"
-      - "baseball bat"
-      - "baseball glove"
-      - "skateboard"
-      - "surfboard"
-      - "tennis racket"
-      - "bottle"
-      - "wine glass"
-      - "cup"
-      - "fork"
-      - "knife"
-      - "spoon"
-      - "bowl"
-      - "banana"
-      - "apple"
-      - "sandwich"
-      - "orange"
-      - "broccoli"
-      - "carrot"
-      - "hot dog"
-      - "pizza"
-      - "donut"
-      - "cake"
-      - "chair"
-      - "sofa"
-      - "pottedplant"
-      - "bed"
-      - "diningtable"
-      - "toilet"
-      - "tvmonitor"
-      - "laptop"
-      - "mouse"
-      - "remote"
-      - "keyboard"
-      - "cell phone"
-      - "microwave"
-      - "oven"
-      - "toaster"
-      - "sink"
-      - "refrigerator"
-      - "book"
-      - "clock"
-      - "vase"
-      - "scissors"
-      - "teddy bear"
-      - "hair drier"
-      - "toothbrush"
