--- src/steelhead_object_recognition/include/steelhead_object_recognition/object_recognizer.hpp
@@ -1,88 +1,0 @@
-#ifndef STEELHEAD_OBJECT_RECOGNITION__OBJECT_RECOGNIZER
-#define STEELHEAD_OBJECT_RECOGNITION__OBJECT_RECOGNIZER
-
-#include "rclcpp/rclcpp.hpp"
-#include "image_transport/image_transport.hpp"
-#include "std_msgs/msg/string.hpp"
-#include "cv_bridge/cv_bridge.h"
-#include "sensor_msgs/image_encodings.hpp"
-#include "steelhead_interfaces/msg/detection_box_array.hpp"
-#include "steelhead_interfaces/srv/object_detection.hpp"
-#include <opencv2/opencv.hpp>
-
-#ifndef DEBUG_VISUALIZE
-    #define DEBUG_VISUALIZE 1
-#endif
-
-namespace steelhead_object_recognition
-{      
-
-    class ObjectRecognizer : public rclcpp::Node
-    {
-
-    public:
-
-        /** Object recognition ROS2 node to detect objects within an image
-         * 
-         * Has a publisher/subscriber and a service to take image inputs and output bounding boxes/classes
-         * 
-         */
-        explicit ObjectRecognizer(const rclcpp::NodeOptions & options);
-
-    private:
-
-        /** Reads image and performs object recognition.
-         * 
-         * Uses the loaded network to perform object recognition on the image.
-         * 
-         * @param msg message containing image data
-         * 
-         * @returns DetectionBoxArray message containing the bounding boxes of detected objects
-         * 
-         */
-        steelhead_interfaces::msg::DetectionBoxArray process(const sensor_msgs::msg::Image & msg) const;
-
-        /** Callback used by subscriber to process and publish results
-         */
-        void subscriberCallback(const sensor_msgs::msg::Image::ConstSharedPtr & msg) const;
-        /** Callback used by service to process and respond to client with results
-         */
-        void serviceCallback(const steelhead_interfaces::srv::ObjectDetection::Request::SharedPtr request, 
-                const steelhead_interfaces::srv::ObjectDetection::Response::SharedPtr response) const;
-
-        rclcpp::Publisher<steelhead_interfaces::msg::DetectionBoxArray>::SharedPtr publisher_;  
-        image_transport::Subscriber subscription_; 
-        rclcpp::Service<steelhead_interfaces::srv::ObjectDetection>::SharedPtr service_;
-        #if DEBUG_VISUALIZE
-            image_transport::Publisher debug_publisher_;
-        #endif
-
-        std::shared_ptr<cv::dnn::Net> net_;
-
-        std::string weights_filename_ = "config/weights/best.pt";  // Path to the PyTorch model
-        cv::dnn::Backend backend_ = cv::dnn::DNN_BACKEND_OPENCV;
-        cv::dnn::Target target_ = cv::dnn::DNN_TARGET_CPU;
-        std::vector<std::string> classes_;
-        float conf_threshold_ = 0.2;
-        float nms_threshold_ = 0.2;
-        float scale_ = 0.00392;
-        int inp_width_ = 608;
-        int inp_height_ = 608;
-        bool swap_rb_ = true;
-        cv::Scalar mean_ = cv::Scalar(0.5, 0.5, 0.5);
-
-        /** Loads image data into network and runs model
-         */
-        void preprocess(const cv::Mat& frame) const;
-        /** Gets relevant data from network output
-         */
-        void postprocess(std::vector<int> & classIds, std::vector<float> & confidences, std::vector<cv::Rect> & boxes, 
-                cv::Mat & frame, const std::vector<cv::Mat>& outs) const;
-    };
-    
-} // namespace steelhead_object_recognition
-
-#include "rclcpp_components/register_node_macro.hpp"
-RCLCPP_COMPONENTS_REGISTER_NODE(steelhead_object_recognition::ObjectRecognizer)
-
-#endif  //STEELHEAD_OBJECT_RECOGNITION__OBJECT_RECOGNIZER
