+++ src/steelhead_object_recognition/package.xml
@@ -3,26 +3,22 @@
 <package format="3">
   <name>steelhead_object_recognition</name>
   <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <description>YOLO object detection / RViz visualization node for Steelhead.</description>
   <maintainer email="kevinh42@student.ubc.ca">kevin</maintainer>
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
+  <buildtool_depend>ament_cmake_python</buildtool_depend>
 
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
-  <depend>std_msgs</depend>
+  <depend>rclpy</depend>
+  <depend>sensor_msgs</depend>
   <depend>steelhead_interfaces</depend>
   <depend>cv_bridge</depend>
-  <depend>sensor_msgs</depend>
-  <depend>image_transport</depend>
-  <depend>realsense2_camera</depend>
+  <depend>python3-opencv</depend>
+  <depend>python3-numpy</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
-  <test_depend>launch</test_depend>
-  <test_depend>launch_testing</test_depend>
-  <test_depend>python3-pytest</test_depend>
 
   <export>
     <build_type>ament_cmake</build_type>
