+++ src/steelhead_object_recognition/setup.sh
